+++ test/forex_currency_conversion_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:forex_conversion/forex_conversion.dart';
+import 'package:forex_currency_conversion/forex_currency_conversion.dart';
 
 void main() {
   test('adds one to input values', () async {