+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.6.5
  - Partial error handling to stop cors related problems on web.
+ - Package rename.
 
 ## 0.6.0
  - Allow the user to initialize the class with default maximum value of decimal places for the currencies.